2024/04/26

+++ workscripts/functions.ps1
@@ -813,6 +813,7 @@ Function Add-MFAExclude {
         `
 
         .EXAMPLE
+        Add-MFAEclude -useremail Brylle.Purificacion@greatgulf.com
         `
     #>
 
@@ -826,24 +827,27 @@ Function Add-MFAExclude {
 
         $userobjectid = Get-MsolUser -SearchString $useremail
 
-        Write-Host
-        Write-Host "Would you like to add the user " -NoNewline
-        Write-Host "'$($userobjectid.DisplayName)' " -ForegroundColor Yellow -NoNewline
-        Write-Host "to the MFA Exclude group? (Y/N): " -NoNewline
-        $proceed = Read-Host
-        Write-Host
+        if ($userobjectid) {
+            Write-Host
+            Write-Host "Would you like to add the user " -NoNewline
+            Write-Host "'$($userobjectid.DisplayName)' " -ForegroundColor Yellow -NoNewline
+            Write-Host "to the MFA Exclude group? (Y/N): " -NoNewline
+            $proceed = Read-Host
+            Write-Host
 
-        if ($proceed -eq 'Y') {
-            Add-MsolGroupMember -GroupObjectId d228403a-1cfd-41b5-a2a4-c5da10d322fa -GroupMemberType User -GroupMemberObjectId $userobjectid.ObjectId
-            Write-Host
-            Write-Host "The user has been successfully added to the 'MFA Exclude' group." -ForegroundColor Yellow
-            Write-Host
-        } else {
-            Write-Host "The operation has been cancelled..."
+            if ($proceed -eq 'Y') {
+                Add-MsolGroupMember -GroupObjectId d228403a-1cfd-41b5-a2a4-c5da10d322fa -GroupMemberType User -GroupMemberObjectId $userobjectid.ObjectId
+                Write-Host
+                Write-Host "The user has been successfully added to the 'MFA Exclude' group." -ForegroundColor Yellow
+                Write-Host
+            } else {
+                Write-Host "The operation has been cancelled..."
+            }
         }
     } catch {
         Write-Host "An error has occurred, please try again later..."
     }
+
 }
 Function Remove-MFAExclude {
     <#
@@ -852,6 +856,7 @@ Function Remove-MFAExclude {
         `
 
         .EXAMPLE
+        Remove-MFAExclude -useremail Brylle.Purificacion@greatgulf.com
         `
     #>
 
@@ -865,19 +870,20 @@ Function Remove-MFAExclude {
 
         $userobjectid = Get-MsolUser -SearchString $useremail
 
-
-        Write-Host
-        Write-Host "Would you like to remove the user " -NoNewline
-        Write-Host "'$($userobjectid.DisplayName)' " -ForegroundColor Yellow -NoNewline
-        Write-Host "to the MFA Exclude group? (Y/N): " -NoNewline
-        $proceed = Read-Host
-        Write-Host
-
-        if ($proceed -eq 'Y') {
-            Remove-MsolGroupMember -GroupObjectId d228403a-1cfd-41b5-a2a4-c5da10d322fa -GroupMemberType User -GroupMemberObjectId $userobjectid.ObjectId
+        if ($userobjectid) {
             Write-Host
-            Write-Host "The user has been successfully removed from the 'MFA Exclude' group." -ForegroundColor Yellow
+            Write-Host "Would you like to remove the user " -NoNewline
+            Write-Host "'$($userobjectid.DisplayName)' " -ForegroundColor Yellow -NoNewline
+            Write-Host "to the MFA Exclude group? (Y/N): " -NoNewline
+            $proceed = Read-Host
             Write-Host
+
+            if ($proceed -eq 'Y') {
+                Remove-MsolGroupMember -GroupObjectId d228403a-1cfd-41b5-a2a4-c5da10d322fa -GroupMemberType User -GroupMemberObjectId $userobjectid.ObjectId
+                Write-Host
+                Write-Host "The user has been successfully removed from the 'MFA Exclude' group." -ForegroundColor Yellow
+                Write-Host
+            }
         } else {
             Write-Host "The operation has been cancelled..."
         }
